requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2-binary==2.9.1
aiogram==2.14.3
requests==2.25.1
apscheduler==3.7.0
